added comments and cleaned up code

+++ config/initializers/redis.rb
@@ -4,4 +4,4 @@ require 'redis'
 
 $redis = Redis.new(Rails.application.config_for(:redis))
 
-$redis.flushdb # if Rails.env == "test"
+$redis.flushdb if Rails.env == "test"
